Add sanitization for given search
I've added a function that should do some sanitization for given searches. The
idea behind it is to make sure that the given PDB ids look like valid PDB ids.
We also will store data about the constraints on each cell if we are given a
size of the search.