Update design
I've learned a bit more about how to structure these sorts of apps from working
on r3d-2-msa. These changes reflect that. In addition, we won't have a python
fr3d before this is finished so do it with a matlab fr3d. This is still mostly a
stub implementation but it is a start.