Several updates for new approach
We will be building both fr3d and webfr3d at the same time. So no need for
matlab stuff. Also the project is being split among several people. This starts
that off by working toward this. We create a basic index template that can be
filled out for the landing or search page.